requirements.txt
aiofiles                  24.1.0
aiogram                   3.14.0
aiohappyeyeballs          2.4.3
aiohttp                   3.10.10
aiosignal                 1.3.1
annotated-types           0.7.0
asyncpg                   0.30.0
attrs                     24.2.0
beautifulsoup4            4.12.3
bs4                       0.0.2
certifi                   2024.8.30
charset-normalizer        3.4.0
click                     8.1.7
colorama                  0.4.6
frozenlist                1.5.0
greenlet                  3.1.1
idna                      3.10
jsonschema                4.23.0
jsonschema-specifications 2024.10.1
magic-filter              1.0.12
multidict                 6.1.0
notifiers                 1.3.3
pip                       24.3.1
propcache                 0.2.0
psycopg2                  2.9.10
pydantic                  2.9.2
pydantic_core             2.23.4
python-dotenv             1.0.1
referencing               0.35.1
requests                  2.32.3
rpds-py                   0.21.0
setuptools                75.3.0
soupsieve                 2.6
SQLAlchemy                2.0.36
typing_extensions         4.12.2
urllib3                   2.2.3
yarl                      1.17.1